seed-experiments/memory-rag/requirements.txt
gradio
tiktoken
langchain
faiss-cpu
sentence-transformers
llama-cpp-python
